Increase column size for node bandwidth_balance

+++ db/pf-schema-X.Y.Z.sql
@@ -154,7 +154,7 @@ CREATE TABLE node (
   unregdate datetime NOT NULL default "0000-00-00 00:00:00",
   lastskip datetime NOT NULL default "0000-00-00 00:00:00",
   time_balance int(10) unsigned DEFAULT NULL,
-  bandwidth_balance int(10) unsigned DEFAULT NULL,
+  bandwidth_balance bigint(20) unsigned DEFAULT NULL,
   status varchar(15) NOT NULL default "unreg",
   user_agent varchar(255) default NULL,
   computername varchar(255) default NULL,
